Move `MujocoCamera` to `mujoco.utils` and define `MujocoCameraType`
